budget-backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
google-auth==2.23.0
google-api-python-client==2.97.0
python-dotenv==1.0.0
pydantic==2.3.0
pyyaml==6.0.1